Tách các giao diện tái sử dụng được
