+++ lib/features/authentication/screens/login/form_divider.dart
@@ -1,0 +1,34 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_store_mobile/utils/constants/colors.dart';
+import 'package:flutter_store_mobile/utils/constants/text_string.dart';
+import 'package:get/get.dart';
+
+class TFormDivider extends StatelessWidget {
+  const TFormDivider({
+    super.key,
+    required this.dark,
+  });
+
+  final bool dark;
+
+  @override
+  Widget build(BuildContext context) {
+    return Row(
+      mainAxisAlignment: MainAxisAlignment.center,
+      children: [
+        Divider(
+            color: dark ? TColors.darkGrey : TColors.grey,
+            thickness: 0.5,
+            indent: 60,
+            endIndent: 5),
+        Text(TTexts.orSignInWith.capitalize!,
+            style: Theme.of(context).textTheme.labelMedium),
+        Divider(
+            color: dark ? TColors.darkGrey : TColors.grey,
+            thickness: 0.5,
+            indent: 60,
+            endIndent: 0.5),
+      ],
+    );
+  }
+}
