+++ lib/features/authentication/screens/login/login.dart
@@ -1,13 +1,11 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_store_mobile/common/style/spacing_styles.dart';
-import 'package:flutter_store_mobile/utils/constants/colors.dart';
-import 'package:flutter_store_mobile/utils/constants/images_string.dart';
+import 'package:flutter_store_mobile/features/authentication/screens/login/form_divider.dart';
+import 'package:flutter_store_mobile/features/authentication/screens/login/login_footer.dart';
+import 'package:flutter_store_mobile/features/authentication/screens/login/login_form.dart';
+import 'package:flutter_store_mobile/features/authentication/screens/login/login_header.dart';
 import 'package:flutter_store_mobile/utils/constants/sizes.dart';
-
-import 'package:flutter_store_mobile/utils/constants/text_string.dart';
 import 'package:flutter_store_mobile/utils/helpers/helper_function.dart';
-import 'package:get/get.dart';
-import 'package:iconsax/iconsax.dart';
 
 class LoginScreen extends StatelessWidget {
   const LoginScreen({super.key});
@@ -24,137 +22,17 @@ class LoginScreen extends StatelessWidget {
           child: Column(
             children: [
               //Logo - sub
-              Column(
-                crossAxisAlignment: CrossAxisAlignment.start,
-                children: [
-                  Image(
-                    height: 150,
-                    image: AssetImage(TImages.appLogo),
-                    //(dark ? TImages.lightAppLogo : TImages.darkAppLogo) dành cho logo chế độ sáng tối//
-                  ),
-                  Text(TTexts.loginTitle,
-                      style: Theme.of(context).textTheme.headlineMedium),
-                  const SizedBox(height: TSizes.sm),
-                  Text(TTexts.loginTitle,
-                      style: Theme.of(context).textTheme.bodyMedium),
-                ],
-              ),
+              const TLoginHeader(),
 
               ///Form
-              Form(
-                child: Padding(
-                  padding: const EdgeInsets.symmetric(
-                      vertical: TSizes.spaceBtwSections),
-                  child: Column(
-                    children: [
-                      //Email
-                      TextFormField(
-                        decoration: const InputDecoration(
-                            prefixIcon: Icon(Iconsax.direct_right),
-                            labelText: TTexts.email),
-                      ),
-                      const SizedBox(height: TSizes.spaceBtwInputFields),
+              const TLoginForm(),
 
-                      //Password
-                      TextFormField(
-                        decoration: const InputDecoration(
-                          prefixIcon: Icon(Iconsax.password_check),
-                          labelText: TTexts.password,
-                          suffixIcon: Icon(Iconsax.eye_slash),
-                        ),
-                      ),
-                      const SizedBox(height: TSizes.spaceBtwInputFields / 2),
-
-                      /// Remember me & Forget password
-                      Row(
-                        mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                        children: [
-                          ///Remeberme
-                          Row(
-                            children: [
-                              Checkbox(value: true, onChanged: (value) {}),
-                              const Text(TTexts.rememberMe)
-                            ],
-                          ),
-
-                          ///Forget password
-                          TextButton(
-                              onPressed: () {},
-                              child: const Text(TTexts.forgetPassword)),
-                        ],
-                      ),
-                      const SizedBox(height: TSizes.spaceBtwSections),
-
-                      //Sign In button
-                      SizedBox(
-                          width: double.infinity,
-                          child: ElevatedButton(
-                              onPressed: () {},
-                              child: const Text(TTexts.signIn))),
-                      const SizedBox(height: TSizes.spaceBtwItems),
-                      //Create account button
-                      SizedBox(
-                          width: double.infinity,
-                          child: OutlinedButton(
-                              onPressed: () {},
-                              child: const Text(TTexts.createAccount))),
-                    ],
-                  ),
-                ),
-              ),
               //Divider
-              Row(
-                mainAxisAlignment: MainAxisAlignment.center,
-                children: [
-                  Divider(
-                      color: dark ? TColors.darkGrey : TColors.grey,
-                      thickness: 0.5,
-                      indent: 60,
-                      endIndent: 5),
-                  Text(TTexts.orSignInWith.capitalize!,
-                      style: Theme.of(context).textTheme.labelMedium),
-                  Divider(
-                      color: dark ? TColors.darkGrey : TColors.grey,
-                      thickness: 0.5,
-                      indent: 60,
-                      endIndent: 0.5),
-                ],
-              ),
+              TFormDivider(dark: dark),
               const SizedBox(height: TSizes.spaceBtwSections),
 
               //footer
-              Row(
-                mainAxisAlignment: MainAxisAlignment.center,
-                children: [
-                  Container(
-                    decoration: BoxDecoration(
-                        border: Border.all(color: TColors.grey),
-                        borderRadius: BorderRadius.circular(100)),
-                    child: IconButton(
-                      onPressed: () {},
-                      icon: const Image(
-                        width: TSizes.iconMd,
-                        height: TSizes.iconMd,
-                        image: AssetImage(TImages.google),
-                      ),
-                    ),
-                  ),
-                  const SizedBox(width: TSizes.spaceBtwItems),
-                  Container(
-                    decoration: BoxDecoration(
-                        border: Border.all(color: TColors.grey),
-                        borderRadius: BorderRadius.circular(100)),
-                    child: IconButton(
-                      onPressed: () {},
-                      icon: const Image(
-                        width: TSizes.iconMd,
-                        height: TSizes.iconMd,
-                        image: AssetImage(TImages.facebook),
-                      ),
-                    ),
-                  ),
-                ],
-              )
+              const TSocialButtons()
             ],
           ),
         ),
