+++ lib/features/authentication/screens/login/login_footer.dart
@@ -1,0 +1,46 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_store_mobile/utils/constants/colors.dart';
+import 'package:flutter_store_mobile/utils/constants/images_string.dart';
+import 'package:flutter_store_mobile/utils/constants/sizes.dart';
+
+class TSocialButtons extends StatelessWidget {
+  const TSocialButtons({
+    super.key,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return Row(
+      mainAxisAlignment: MainAxisAlignment.center,
+      children: [
+        Container(
+          decoration: BoxDecoration(
+              border: Border.all(color: TColors.grey),
+              borderRadius: BorderRadius.circular(100)),
+          child: IconButton(
+            onPressed: () {},
+            icon: const Image(
+              width: TSizes.iconMd,
+              height: TSizes.iconMd,
+              image: AssetImage(TImages.google),
+            ),
+          ),
+        ),
+        const SizedBox(width: TSizes.spaceBtwItems),
+        Container(
+          decoration: BoxDecoration(
+              border: Border.all(color: TColors.grey),
+              borderRadius: BorderRadius.circular(100)),
+          child: IconButton(
+            onPressed: () {},
+            icon: const Image(
+              width: TSizes.iconMd,
+              height: TSizes.iconMd,
+              image: AssetImage(TImages.facebook),
+            ),
+          ),
+        ),
+      ],
+    );
+  }
+}
