+++ lib/features/authentication/screens/login/login_header.dart
@@ -1,0 +1,28 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_store_mobile/utils/constants/images_string.dart';
+import 'package:flutter_store_mobile/utils/constants/sizes.dart';
+import 'package:flutter_store_mobile/utils/constants/text_string.dart';
+
+class TLoginHeader extends StatelessWidget {
+  const TLoginHeader({
+    super.key,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return Column(
+      crossAxisAlignment: CrossAxisAlignment.start,
+      children: [
+        Image(
+          height: 150,
+          image: AssetImage(TImages.appLogo),
+          //(dark ? TImages.lightAppLogo : TImages.darkAppLogo) dành cho logo chế độ sáng tối//
+        ),
+        Text(TTexts.loginTitle,
+            style: Theme.of(context).textTheme.headlineMedium),
+        const SizedBox(height: TSizes.sm),
+        Text(TTexts.loginTitle, style: Theme.of(context).textTheme.bodyMedium),
+      ],
+    );
+  }
+}
